+++ common/src/main/java/de/cech12/ceramicbucket/init/ModTags.java
@@ -1,10 +1,10 @@
 package de.cech12.ceramicbucket.init;
 
-import de.cech12.ceramicbucket.CeramicBucketMod;
+import de.cech12.ceramicbucket.Constants;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.level.material.Fluid;
-import net.minecraftforge.registries.ForgeRegistries;
 
 import javax.annotation.Nonnull;
 
@@ -15,7 +15,7 @@ public class ModTags {
         public static final TagKey<Fluid> CERAMIC_CRACKING = tag("ceramic_cracking");
 
         private static TagKey<Fluid> tag(@Nonnull String name) {
-            return TagKey.create(ForgeRegistries.FLUIDS.getRegistryKey(), new ResourceLocation(CeramicBucketMod.MOD_ID, name));
+            return TagKey.create(Registries.FLUID, new ResourceLocation(Constants.MOD_ID, name));
         }
 
     }